+++ src/app/app.module.ts
@@ -1,6 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { Router } from '@angular/router';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import { ProductsModule } from './products/products.module';
 import { CartModule } from './cart/cart.module';
@@ -11,11 +12,13 @@ import { SharedModule } from './shared/shared.module';
 import { AppRoutingModule } from './app-routing.module';
 
 import { AppComponent } from './app.component';
+import { TimingInterceptor } from './core/interceptors/timing.interceptor';
 
 @NgModule({
   declarations: [AppComponent],
   imports: [
     BrowserModule,
+    HttpClientModule,
 
     ProductsModule,
     CartModule,
@@ -25,6 +28,13 @@ import { AppComponent } from './app.component';
 
     // LAST ONE
     AppRoutingModule,
+  ],
+  providers: [
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: TimingInterceptor,
+      multi: true,
+    }
   ],
   bootstrap: [AppComponent],
 })