+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,7 +1,11 @@
 import { Component, OnInit } from '@angular/core';
-import { ProductsService } from 'src/app/products/services/products.service';
-import { CartService } from 'src/app/cart/services/cart.service';
 import { Router } from '@angular/router';
+
+import { Observable } from 'rxjs';
+
+import { CartService } from 'src/app/cart/services/cart.service';
+import { ProductHttpService } from '../../services';
+import { ProductModel } from '../../model/product.model';
 
 @Component({
   selector: 'app-product-list',
@@ -9,17 +13,18 @@ import { Router } from '@angular/router';
   styleUrls: ['./product-list.component.scss'],
 })
 export class ProductListComponent implements OnInit {
-  productList: Promise<Item[]>;
+  productList: Observable<ProductModel[]>;
+  // : Promise<ProductModel[]>;
 
   constructor(
-    public productsService: ProductsService,
     public cartService: CartService,
     private router: Router,
+    private productHttpService: ProductHttpService,
   ) {}
 
   ngOnInit() {
-    this.productList = this.productsService.getProducts();
-    this.productsService.isDisplayed = false;
+    this.productList = this.productHttpService.getProducts();
+    this.productHttpService.isDisplayed = false;
   }
 
   onBuy(item: Item): void {
@@ -35,5 +40,15 @@ export class ProductListComponent implements OnInit {
   onEditProduct(item: Item): void {
     const link = ['product/edit', item.id];
     this.router.navigate(link);
+  }
+
+  onDelete(product: ProductModel): void {
+    this.productList = this.productHttpService.deleteProduct(product);
+
+  // for service with Promise
+  //   this.productHttpService
+  //     .deleteProduct(product)
+  //     .then(() => (this.productList = this.productHttpService.getProducts()))
+  //     .catch(error => console.log(error));
   }
 }